Search/requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2
django-filter==2.3.0
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.6.1
Markdown==3.2.2
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
zipp==3.1.0
